Re-hook the Google Map callbacks on iOS. (#934)
The refactoring in #892 left the callbacks defined in `FLTGoogleMapDelegate` unhooked.
Fixes flutter/flutter#24520.

This change also removes the `FLTGoogleMapDelegate` protocol, it was used to
delegate map callbacks from the map controller to `FLTGoogleMapsPlugin`
which was the owner of the only method channel. As we now have a method
channel per map controller there is no need for a delegate.

This also renames `didTapInfoWindow` to `didTapInfoWindowForMarker` in
`FLTGoogleMapController`. This was just a typo when the code was first
introduced, and fixes flutter/flutter#20178.

Additionally this change cleans up the no longer used `addToView` and
`removeFromView` method from `FLTGoogleMapController`. These were
used by the overlay-based implementation, in the `UiKitView`
implementation the view is added and removed by the Flutter Engine.