[google_maps_flutter] Avoid AbstractMethod crash (#2113)
Override a default method to work around flutter/flutter#40126. The root
cause of the issue has already been fixed in master and there is a
potential workaround available to apps already, but this is an extra
low-risk change that should prevent the crash for any affected users
while the fix is still rolling out to stable.